add get note by user login and authenticate link

+++ app/Http/Controllers/NoteController.php
@@ -12,7 +12,10 @@ class NoteController extends Controller
      */
     public function index()
     {
-        $notes = Note::query()->orderBy('created_at', 'desc')->paginate();
+        $notes = Note::query()
+        ->where('user_id', request()->user()->id)
+        ->orderBy('created_at', 'desc')
+        ->paginate();
         return view('note.index', ['notes' => $notes]);
     }
 
@@ -33,7 +36,7 @@ class NoteController extends Controller
             'note'=>['required', 'string']
         ]);
 
-        $data['user_id'] = 1;
+        $data['user_id'] = $request->user()->id;
 
         $note = Note::create($data);
 
@@ -45,6 +48,9 @@ class NoteController extends Controller
      */
     public function show(Note $note)
     {
+        if ($note->user_id != request()->user()->id) {
+            abort(403);
+        }
         return view('note.show', ['note' => $note]);
     }
 
@@ -53,6 +59,9 @@ class NoteController extends Controller
      */
     public function edit(Note $note)
     {
+        if ($note->user_id != request()->user()->id) {
+            abort(403);
+        }
         return view('note.edit', ['note' => $note]);
     }
 
@@ -61,6 +70,10 @@ class NoteController extends Controller
      */
     public function update(Request $request, Note $note)
     {
+        if ($note->user_id != request()->user()->id) {
+            abort(403);
+        }
+
         $data = $request->validate([
             'note'=>['required', 'string']
         ]);
@@ -75,6 +88,9 @@ class NoteController extends Controller
      */
     public function destroy(Note $note)
     {
+        if ($note->user_id != request()->user()->id) {
+            abort(403);
+        }
         $note->delete();
 
         return to_route('note.index')->with('message', 'Note was deleted');
